code/project-template/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["haruiz <henryruiz22@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
numpy = "^1.26.4"
scikit-learn = "^1.3.0"
python-multipart = "^0.0.9"
pillow = "^10.2.0"


[tool.poetry.group.tf_mac]
optional = true

[tool.poetry.group.tf_mac.dependencies]
tensorflow-macos = "^2.13.0"
tensorflow-metal = "^1.0.1"

[tool.poetry.group.tf_win]
optional = true

[tool.poetry.group.tf_win.dependencies]
tensorflow = "^2.13.0"

[tool.poetry.group.tf_linux]
optional = true
[tool.poetry.group.tf_linux.dependencies]
tensorflow = "^2.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
